Adding Vehicle Types Pages

+++ server/BusinessModels.Abstractions/Vehicles/Types/Commands/IUpdateVehicleTypeDTO.cs
@@ -6,5 +6,5 @@ public interface IUpdateVehicleTypeRequirement
     : IUpdateEntityCommandRequirement
 {
     string Name { get; }
-    int NumberOfPassengers { get; }
+    long NumberOfPassengers { get; }
 }
